Present messages when archiving emails
Previously we only (optionally) presented the content change associated
with an archived email. This extends the presenter to also cope with
emails originating from ad-hoc messages. Since content changes and
messages are the only potential origins, I've gone with a duplication
approach in the code/tests, rather than trying to abstract anything.

One of the tests, which checks exception notification behaviour, seemed
unnecessary, since it's tangential to the behaviour of the presenter.
Otherwise, coping with an array of digest IDs should still be covered
implicitly by the use of an array in the input records.