Record timing of full email batch inserts
This sets up metrics to record when a full batch of 5000 emails is
inserted into the database. We only want to time a full batch as that
provide an accurate basis for performance comparison. The intention is
to use this data to get an understanding of the performance impacts of
these large inserts.

The statsd key uses the number of items inserted so that any changes to
this number can be represented as a new data point.