Remove the Subscription Contents healthcheck
The `email-alert-api app healthcheck not ok` icinga alert often
fails due to a timeour error because some of the checks are slow.

We are removing this healthcheck in favour of a Sidekiq job and
Icinga check. This will allow us to send data to Graphite so that
we can monitor the number of subscription contents with `critical`
and `warning` latency.

See hhttps://github.com/alphagov/email-alert-api/pull/1096 and
https://github.com/alphagov/govuk-puppet/pull/10017 for more
information.