Record stats on subscription contents
We are moving away from using healthchecks to monitor specific
things as they were putting unnecessary load on the database on
each machine and triggering duplicate Icinga alerts.

This sidekiq job will run every 1 min and send the total number
of `critical` and `warning` subscription contents to Graphite.
We can use StatD's gauge feature to maintain the current values
until they are next updated.

We can then configure an Icinga alert in `Puppet` which will trigger
when certain values are reached. This will not longer be machine
specific and come through from the monitoring machine.