Record stats on digest runs
We are moving away from healthcheck monitoring for individual
metrics as this places additional DB load on each machine and
creates duplicated Icinga alerts.

This PR adds a sidekiq job which to send critical and warning
digest run metrics to Graphite. This is in advance of a new
Icinga alert being configured in Puppet to monitor these values
and report.