Remove Messages from healthcheck
The `email-alert-api app healthcheck not ok` icinga alert often
fails due to a timeour error because some of the checks are slow.

We are removing this healthcheck in favour of a Sidekiq job and
Icinga check. This will allow us to send data to Graphite so that
we can monitor the number of messages with `critical` and `warning`
latency.