Switch to default sidekiq_retry_in
Sidekiq has a reasonable default retry delay that implements an
exponential back off meaning that the jobs will take longer and longer
to retry.

I believe that we originally set this up in the way we did because we
were concerned that our internal rate limit might mean that emails get
delayed exponentially when the system only needed to wait a few seconds.
However, in reality it appears as though our internal rate limiting
system almost never gets triggered, and a more likely scenario is that
Notify could have problems on their side. In this case, we don't want to
try again so soon (10 seconds as it is currently configured) but instead
we'd want to try longer and longer periods until the problem on Notify's
side clears up.