Increase number of delivery retry attempts
Currently it's set to 3 which means that emails are only attempted
within one minute before failing with a technical failure. If instead we
set it to 9, with the default exponential back off time, Sidekiq would
try the job at the longest 90 minutes after the first attempt.

Recently Notify had an incident that lastest 15 minutes, and because of
that most of the delivery attempts in that time period failed and never
succeeded after retrying. We had to manually intervene, whereas if we
configure Sidekiq to try for a longer period of time, it should cope
with a Notify incident.