Use find_by_address method in unsubscribe rake task
Currently we are matching on a case sensitive email address when
manually unsubscribing users.  Changing to use the find_by_address
method to match behaviour in the UI.

This has caused an issue where a developer unsubscribed an email
containing an uppercase character, but the user's email was stored all
lowercase.  This meant they were never unsubscribed.