Remove top-level rake task methods
This fixes a linting issue for top-level methods in rake tasks by
moving the definitions of those methods into the rake task that used
them. In the case of moving subscribers between lists, the resulting
rake task was then too long (new linting issue), so I moved the code
into a new class, which makes sense for such a large volume of code.