Mark emails that haven't sent because of a technical failure as "failed"
Previously, if a delivery attempt was marked as "technical_failure",
the email would stay marked as "pending".

This meant we weren't able to use [this rake task][rake-task] to resend
failed emails during a recent incident (after the cause of the failure
had been fixed), so we had to manually query for them.

Trello: https://trello.com/c/M8dulgoy/1938-3-mark-emails-that-havent-sent-because-of-a-technical-failure-as-failed

[rake-task]: https://github.com/alphagov/email-alert-api/blob/c7c03a0f292df40d4d7c2957014b67d8ac11b2dc/lib/tasks/deliver.rake#L24-L32