Create 'view_subscriptions' rake task
This can be used in tandem with the 'unsubscribe_single_subscription'
and 'unsubscribe_single' rake tasks so that we can double-check
a user's subscriptions both before and after applying the updates.

This new rake task facility should mean we rarely need to physically
SSH into email-alert-api machines anymore.

The custom table output takes inspiration from
https://stackoverflow.com/questions/28684598/print-an-array-into-a-table-in-ruby
I could have added a gem dependency to delegate columnised output
to a third party, but writing a bit of code seems reasonable given
that it's only used in isolation in this one rake task.