Fix email archiving on technical failure
When emails are marked as technical failure they do not have a
finished_sending_at time. This means that the EmailArchivePresenter
fails later when trying to read the field. Here we have changed the
method which updates the status of the delivery_attempt to set the
finished_sending_at time to the current time so the field is not nil.

Trello card: https://trello.com/c/nbNAVLwS/1965-3-fix-archiving-of-technical-failure-emails