Fix sending emails to migrated subscribers
The rake task for migrating subscribers to a different subscriber list
sent emails before the migration was completed. Especially for large
lists of users, this caused a confusing experience as emails were
received while the migration was still happening.

This change allows to to generate the emails to send, make the
migration, and then send the emails after the migration has
completed.