Add a metric to record first_delivery_attempt
This change is rather bonkers as it adds a metric to record how long it
takes to record another metric in some sort of weird metric-ception.
From digging into where the time is going when trying to send emails the
queries involved in pulling out this metric appear to be where the
majority of time goes when trying to send an email.

This is being recorded in the hope it helps fill in our picture of where
the time goes when sending emails. It's a bit unclear at this stage what
we're going to do to resolve this lost time.