Move database queries outside the MetricsService
This caught us out when investigating where time was spent in trying to
understand Email Alert API performance as we were surprised to learn the
innocuously named MetricsService was actually talking to the database.

This moves all of the metric calls that use the DB into the
DeliveryRequestService so they can be identified easier.