Remove import_ignoring_duplicates
Switch to using import! from the activerecord-import gem. We're now
using PostgreSQL 9.6, so we should be able to use the gem directly, as
described in the code comment.

I've updated the tests for this area, just to help demonstrate that
the behaviour is consistent. Maybe the tests are irrelevant now.