Create ImmediateEmailGenerationService
This class, and it's nested Batch class, are intended to be used to
create all immediate emails for a ContentChange or Message and then put
them on the queue for delivery.

It is intended to be used by the
ProcessContentChangeAndGenerateEmailsWorker (and message equivalent)
to replace looking this information up within the model.

It offers a simpler approach to generating these emails to what we currently
have. Our current approach is based on a time when we had a two step
process where processing a content change meant populating the
SubscriptionContent table and a later process read that table to create
the emails. This new approach switches to us keeping the
content change, the subscribers and subscriptions in memory and thus
being able to insert emails and SubscriptionContent in the same
transaction.

I did consider whether to have individual classes for ContentChange and
Message or use inheritance (as is currently used). In the end I concluded
the overhead of dealing with both content change and message isn't too
high in here so is probably easier digested without
inheritance/composition

These class are expected to be able to deal with a couple of somewhat
unusual situations to allow them to be resilient to problems that could
occur where a ContentChange/Message processing aborts prematurely or if users
subscriptions change during the task duration (which may span a number
of minutes). I'm not sure if some of these are a little overly cautious,
but I don't think they offer performance detriments. Anyway, they are:

- When processing a batch we check to see if subscription_contents
  already exist and filter them out. This is done in the expectation
  that a processing blew up mid way through so some emails were created
  but not all
- Looking up Subscription.immediately.active to only include immediate
  active subscriptions since they job initially started, this is
  consistent with [1]
- Looking up Subscriber.activated in case the subscriber is no longer
  activated since we started the job however many minutes ago (this is
  consistent with [2], I'm not sure how necessary this is given we check
  subscription is active and these should be consistent.

[1]: https://github.com/alphagov/email-alert-api/blob/317523aa6cc6a210d1b13b61a6cbbe075be9d939/app/workers/process_content_change_and_generate_emails_worker.rb#L41-L49
[2]: https://github.com/alphagov/email-alert-api/blob/317523aa6cc6a210d1b13b61a6cbbe075be9d939/app/queries/subscribers_for_immediate_email_query.rb#L16