Only stub $stdout where we need to
Previously we swallowed calls to 'puts' in any test, which created
confusion for people trying to use an interactive debugger. Since
RuboCop prevents uses of 'puts' outside of 'lib/', it should not be
necessary to swallow calls to this method in general.

This removes the stub and applies it in the specific places where it's
necessary to avoid noisy test output. Note that some of the calls to
'puts' were due to the test running the code twice, with the instance
correctly swallowing the output already.