Don't re-match content changes/messages
This applies a guard clause to both of these services to prevent items
being re-imported when they already have been done. This is because the
workers related to these jobs don't run in transactions so could be
resuming partially through an existing processing. If this is the case
we can avoid the need to re pull this information out and repopulate the
database.