Use ActiveRecord#insert_all for generating emails
This replaces the use of the activerecord-import gem for the bulk
inserts of Email and SubscriptionContent in the process of generating
emails.

I've switched the places where we call `#ids` on returned from functions
to instead return an array of ids to save this extra call.

I've also had to add the guard clause to check whether we have passed
the builder an empty collection. I'm going to see if that can be removed
in a subsequent PR.

Since Rails 6 ActiveRecord has provided it's own bulk insert
implementation so we now no longer need to use the gem. The Rails
implementation is at at a slightly lower level which means we need to
provide data such as created_at/updated_at timestamps. Hopefully this
lower level also provides an increase in performance.