Use ActiveRecord#insert_all for importing subscription_contents
This replaces the use of the activerecord-import gem with the insert_all
methods that were provided in Rails 6. The ActiveRecord approach is
lower level so needs timestamps applied directly, it also won't run
validations. I don't forsee the latter being a problem in our code
because we don't base it's input on any user input and only use
the validation for a developer protection. This seems somewhat wasteful
given it will mean creating millions of models to check this validation
on each day.

The use of #insert_all! instead of #insert_all is to raise exceptions if
any duplicate key constraints are hit. With #insert_all these duplicate
keys will be ignored.