Use ActiveRecord#insert_all! on MatchedContentChange/MatchedMessage
This switches both of these classes to use the functionality provided by
ActiveRecord.

In theory it wouldn't be too big a problem if these used the non-bang
variation of the method and ignored duplicate keys. However this seems
to create an impression that we think it'd be ok for two threads to be
trying to match content changes at the same time and I think this is a
confusing situation that we should avoid. It seems likely we'll be
adding some locking to prevent that occurring.