Remove the SubscriptionContent presence validation
This validation no longer has an effect since switching to use
ActiveRecord#insert_all to insert records. It's a bit annoying to lose
this as this is used to maintain a degree of database consistency.

Ideally, we'd replace this with a PostgreSQL constraint check, for
example: `CHECK (num_nonnulls(content_change_id, message_id) = 1);` [1]
however we can't apply a PostgreSQL constraint without switching from
the regular db/schema.rb to db/schema.sql which is less conventional and
much harder to work with (for example it specifies db version which
makes it difficult to upgrade the db in a backwards compatible way).

I'm thinking the likelihood of this check to be needed is low and it's
mostly just a safety net for code mistakes, I'm thinking we can probably
survive without it. If it becomes a problem we could always add a gem
[2] if we feel we need it.

[1]: https://stackoverflow.com/a/55223096
[2]: https://github.com/on-site/active_record-postgres-constraints