Living In [Europe] task uses export_detailed_csv
The FCO have requested detailed information about the Living In guides, including:

New subscribes
New unsubscribes
Immediately
Daily
Weekly
Total sign-ups

The export_detailed_csv provides this out of the box, so switch to using it. This also means we can now provide a date to the rake task and it will get data for that date.