Add documentation on queues
During incidents, it's often not clear to people what the various queues
are used for (one specific example is the difference between `immediate`
and `immediate_high_priority`).

This adds documentation on each of the queues, which will get picked up
by the developer docs.

This work came from an incident we had recently related to the system
being heavily overloaded.