Consolidate and reprioritise similar queues
Previously we had a different weight for each queue, which only
leads to small differences; we also prioritised delivering emails
over generating them, which doesn't make sense, since jobs to deliver
email are far more numerous. This reprioritises the queues so that
queues for similar operations have the same weight, which means they
are checked at the same frequency; we use powers of 2, so that higher
priority queues are checked twice as much as the previous group.

This also increases the processing frequency for processing changes,
so that we prioritise generate emails over sending them.