Remove unnecessary validation for DigestRunSubscriber
These fields are required at a db level and this model is not created as
the result of user input. These validations will no longer run since we
use Rails' #insert_all method to populate this table. This shouldn't be
a loss though as these validations would have merely been an exception
that was raised prior to PostgreSQL raising it's own error.