Add method to bulk populate SubscriptionContent
This method acts as a helper method for the insert_all function allowing
some of the complexities needed for dealing with that to be abstracted
away.

It is intended to be used to populate all of the subscription content
records for a content change or message when emails have been created.
This is different to the current approach where subscription content is
populated and then iterated to create emails. Because of this difference
this is switched to raise errors if it encounters a duplicate
subscription content, the previous approach of ignoring them would mean
that we end up with duplicate email records.

It is named populate_for_content to reflect that there will likely be a
partner method of populate_for_digest_run_subscriber to handle the other
main case for populating this table.