Replace Sidekiq queue weights with priorities
Previously we had an incompatible mixture of ordering [1] and weights,
which were often set arbitrarily and misread as priorities [2]. This
removes all the weights, so that the configuration simply prioritises
the queues based on the order they are listed.

This also groups similar queues together: for delivery and generation;
we should prioritise digest email above cleanup and 'default' jobs.

[1]: 9473c38a4e7eb989cc39c84fd15fe9e0ffc4c136
[2]: 21dcb06850bb6ccbe857bc04d05f5b476bba96ea