Combine metric collection workers into one
Previously we had a 'X' Worker for each model we wanted to collect
periodic metrics about. These names were confusing because they implied
some actual processing associated with 'X', and introduced noise into
the list of workers. This moves the code to a new 'Metrics' namespace
in 'lib/', and invokes each 'Exporter' from a single worker.

Note that some of the tests have been simplified to remove unnecessary
shared_examples and contexts, which appear to be a result of copy/paste.
I've also removed all tests for the queue used by the worker, since
that's declarative and there's little risk in not testing it.