Inline methods from GlobalMetricsService
Previously this class was a layer of abstraction between the classes
generating the metrics, and the mechanism of sending them to Statsd.
Since this is the sole purpose of the classes generating the metrics,
and the mechanism is trivial and single-use (per metric), it's much
clearer to inline the sending of the metrics alongside their collection.

Note that the Statsd configuration used by GlobalMetricsService matches
the defaults for GovukStatsd.