Increase rate limit to 21,000
We previously temporarily reduced the rate limit to give Notify more breathing space to an investigate an issue they were seeing.
They've given us the go-ahead to increase again.

Our original limit was 360/s (21,600). This was weirdly just above Notify's limit, when really we want it to be just under or the same. Setting to 350/s for now so it's the same as Notify's. We can chat about getting Notify to increase things on their side in the future once they've confirmed the issue they were seeing has been resolved.