Make Gemfile consistent
This removes all the various types of gem ordering and version pinning
across this file to be consistent.

I've gone with an approach of specifying no versions at all since all of
these gems are on the latest versions. I'm wondering about trying to
specify/start a convention on GOV.UK gemfiles where we only specify the
version in the Gemfile when we have a reason to pin and requiring a
comment when we do specify to indicate why.