The most modest refactor of create/delivery feature specs
So my goal in these files was to switch from Timecop to
ActiveSupport::Testing::TimeHelpers. This was rather tricky as Timecop
was used in a nested scenario and, wow, this was one long test with a lot
of things going on.

I've tried a couple of very modest tweaks to try make these a bit easier
to understand. I've split it into two separate tests and moved them
around a bit.

I think the approach to testing here needs a re-think to sort these out
properly. It seems to be a feature test at a unit test level of
fidelity and that fidelity should probably be toned down a bunch.