Nest and name RateLimitExceededError correctly
This class was previously in the global namespace, which was
unconventional and it seemed to have a typo in the name since we
separate rate and limit as separate words in the code.

This moves this error into the class, changes it's superclass to the
more appropriate RuntimeError (after all this is an error that occurs
due to runtime conditions) and changes the naming.

This then adds testing for the retries exhausted behaviour since I
wasn't sure how that might need to change and felt it was a blind spot
in our testing strategy.