Refactor DeliveryRequestWorker and associated spec
This applies some basic refactoring to this class to remove the unused
instance variables, to mark only privately used methods appropriately
and replace the check_rate_limit! with a non-bang version (as per [1])

This does similar to the tests, where the tests for the semi-private
methods are removed. There are a few other minor tweaks to make them
more accurate.

[1]: https://rubystyle.guide/#dangerous-method-bang