Move EmailAddressOverrider to separate file
Since this file has it's own unit tests it doesn't appear to be a
private class, thus it makes sense to be a file in it's own right.

I did debate about whether to modernise the tests somewhat but this
whole override system seems somewhat crazy (see config/email_service.yml
for further evidence) so felt it was best to just copy verbatim.