Refactor DeliveryRequestService and spec
This is a change to make this class act in a more conventional fashion
for a service with a single call method by moving the arguments all into
the initialize block.

Aside from that I've made email into an instance variable, since it's
used a lot, and renamed reference to be attempt_id since that seems to
better capture it's usage. I've renamed the call_provider method to
send_email to better capture the intent and moved the provider look up
into there since it's only used once in the class. I've also set address
as a lazy instance variable so that can encapsulated in a single method.

I've done a full refactor of the tests to reflect that the only public
method is .call and that everything should be tested through that. In
the process I dropped some of the metrics tests since their logic was
trivial and we weren't testing any side effects of the code.