Replace queries for metrics with injection
Looking up the data of when a content change was created takes two
queries, one of these plans rather badly (looking up the
SubscriptionContent for an immediate subscriber). This lookup can be
quite spiky in our metrics taking up to a second.

This removes the queries in place of injecting the metrics in directly
from the job to avoid any need to look up the data.