Accept a 2nd parameter of metrics for DeliveryRequestWorker
This is intended to be used to pass any data into the job that may be
used to record information about the job. This is currently limited to
when a content change is created at and may only ever be this (though a
rather clear omission is the lack of an equivalent of message).

There is some temporary code in here to allow for the second argument to
be handled if it is not a hash, this is included as current jobs in the
queue have a second argument of the queue name which is not a hash.

As Sidekiq populates arguments by parsing JSON we are limited in
idiomatic approaches to defining the arguments. To try make this clear
I've added an explicit parsed_metrics method to apply a conversion to a
symbol based hash with a time object.