Pass metrics to the DeliveryRequestWorker
This injects a hash of metrics to the DeliveryRequestWorker so it
doesn't need to look up these values from the DB. This only includes the
time content change was created at.

I've used a memoised variable since this same value may be looked up
over 100,000 times during this codes execution.