Create 'rake manage:view_emails' task for troubleshooting
It can be fiddly to try to retrieve all email delivery attempts made to
a given user, in the event that they complain that they have not received
an email. Especially so when considering "Confirm your subscription"
emails, which aren't yet associated with a subscription.

This rake task attempts to automate some of that work.