Increase latency time required for alert
Currently we're in a situation where these warning / critical
thresholds are too low and our alerting system is raising these alerts
pretty frequently and causing quite a lot of noise for 2nd line.

We want to increase these latency thresholds to a point where
they're actually alerting when something is wrong rather than when
email-alert-api is processing a large batch of emails.

These increases also align with the work we're doing to automatically
requeue content changes and messages if they haven't been processed
within 1 hour. Once this is implemented we want to give email-alert-api
a chance to sort it self out before these alert and manual intervention
is needed. The re-queue PR is [here](https://github.com/alphagov/email-alert-api/pull/1318).

Trello card:
https://trello.com/c/Li8Zo03f/381-relax-the-thresholds-for-content-change-and-message-alerts