Add rake task to fix /world email subscriptions
In an attempt to migrate world location email signups out
of Whitehall, a number of subscriptions became corrupted [1].
While we've stopped the issue for new subscriptions [2], the
corrupted ones remain and we need to fix them.

Original pages:

1. /world/spain (world location taxon)
2. /world/living-in-spain (world location child taxons)
3. /world/spain/news (world location)

Corruption:

- 1 + 3 => "4" (impossible to untangle)
- 2 => 2a (possible to auto-fix)

In order to cease the user impact and officially close the
incident, we have agreed there are two things for us to do:

1. Report the number of subscriptions for each corrupted list.
This can just be the subscriber list title, with the active subscription count.

2. Write a rake task (to be safe) to delete "4"-type subscriptions,
and auto-fix "2a"-type subscriptions.

The ~1300 corrupted lists we are interested in have links like:

```
"world_locations"=>{"any"=>["<content_id>"]}
```

We need to disambiguate between "4"-type and "2a"-type lists,
based on the Content ID in their links:

- "2a"-type Content IDs will match a level one world location
taxon with the same slug as one of the world locations in Whitehall.*

- "4"-type Content IDs will match any other world "taxon" Content
ID. We can get the full list of world taxons from the Publishing API.

*the slugs are the same because world location taxons previously
took over from world locations in Whitehall, in a failed attempt to
replace them - they continue to coexist, with different purposes.

[1]: https://docs.google.com/document/d/1fL5mQjRX2ApoVd4rOIxrnwjhrC-w4pgNgNXaA93KeM8/edit
[2]: https://github.com/alphagov/email-alert-frontend/pull/714