Remove the failure_reason column from Email
This is a column that became rather confused once it had
technical_failure added to it. It opened the question of how does
technical_failure differ from a retries_exhausted_failure and whether
they are permanent failure or not.

Since our usage of this column is near non-existent it seems that the
simplest course of action would be to remove it from the application.
We'll continue to have a status available in a delivery attempt which
could always be used to determine a failure reason or to backfill this
if we want to re-introduce this column.

In order to apply this without causing an application crash it is
prudent to start by ignoring the column and in a later deploy drop it.