Change the DeliveryAttempt status values
This reflects a desire to move away from the coupling, and associated
confusion, of mixing Notify terminology with our own.

The new failure statuses mean the following things:

- undeliverable_failure: we can't deliver email to this users account,
  this could because of an invalid address, an account that doesn't
  exist anymore, a full mailbox or some other complication outside this
  apps control. This combines the previous statuses of permanent_failure
  and temporary_failure
- provider_communication_failure: Our attempt to communicate with the
  provider failed for some reason. For example the HTTP request failed
  for some reason. This most closely represents the past status of
  technical_failure as this was almost exclusively used for situations
  when we couldn't send.

I decided for ease of rollback and for relevancy to previous data it was
best that undeliverable_failure replaced temporary_failure and
provider_communication_failure replaced technical_failure. This does
leave the enum digits looking a little awkward for skipping a number but
the integers aren't supposed to mean anything.

Any existing DeliveryAttempt objects that have a value outside these
status values will load with a status of nil. I don't expect this will
be a problem in practice as the DeliveryAttempt status isn't read
outside of the point of writing the value.