Stop sending spam data to S3/Athena
We currently send `marked_as_spam` stats to Athena but they're not actually visible
and have dubious accuracy if we receive a spam report _after_ the one-time action of
sending the archive data to Athena. We should stop sending this data, in favour of
using the metrics added as part of this pull request.