Stop retrying emails that Notify fails to send
This makes a change to StatusUpdateService to no longer re-attempt to
send emails that experience a temporary failure. The reason for this is
because these make up a very small proportion of emails we send and the
main reason for this to occur is "mailbox full" which feels like a user
error outside of our control.

I looked into the data on 23rd July and found that over the last 7 days
we had sent 24 million emails, of those only 14,175 had experienced a
temporary-failure, which equates to 0.05% of emails being sent as a
result of this logic. We found that over the same period 122,122 emails
experienced a temporary-failure which represents 0.5% of emails sent. By
comparing the 122,122 temporary failure emails with the 14,175 that were
successful we can determine that 89% of Emails that experience a
temporary-failure are never successful. Therefore, this indicates that the
extra work we do to retry these emails is only a benefit to an incredibly
small segment of the audience.

As part of updating this service I've also changed the following items:

- replaced the need for UpdateEmailStatusService, this wasn't necessary
  once the complexity of retrying was removed;
- stopped handling technical-failure responses from Notify, from a Slack
  chat with Notify on July 21st 2020 we learnt that Notify does not send
  callbacks for technical-failure statuses, thus this code is redundant;
- stopped recording when StatusUpdateService fails, we can monitor HTTP
  responses for this information;
- record the status of a status update status from Notify in Statsd, this
  allows replacing status_update.success with more specific information
  that reflects that delivery_attempt status is no longer equal in
  information;
- writes to the Rails log when we experience an unsuccessful Notify
  response; this is to provide debugging ability now this information is
  no longer stored in the database.