Actually fail emails after retries
Previously when Sidekiq ran out of retries for delivery requests it
would stop trying to send an email without recording any change in
status. This led to Email objects that had a perpetual status of pending
yet nothing was attempting to send them. This results in them being left
in the database forever.

Instead this marks an email as failed after we have exhausted all of the
attempts to send it. The expectation is that 9 attempts is sufficient to
rule out a transient failure and represent a systemic one. These 9
attempts will use Sidekiq's default retry back-off algorithm [1] which
roughly equates to 2-3 hours from first to last attempt.

[1]: https://github.com/mperham/sidekiq/wiki/Error-Handling#automatic-job-retry