Stop sending Notify error responses to Sentry
We experience a number of transient errors during our regular operations
with Notify that don't indicate an actionable problem to solve. Common
ones are:

- Notifications::Client::ServerError - we receive this mostly for a 502
  response, but also semi-regularly see 500 errors
- Notifications::Client::AuthError - we see this error a lot for
  "AuthError: Error: Your system clock must be accurate to within 30
  seconds", this however is un-actionable as Notify took over 30 seconds
  to handle the request
- Notifications::Client::BadRequestError - we see this quite regularly
  when we request sending an email with an invalid address, we also see
  this on occasions in integration and staging where we see an error of
  "Can’t send to this recipient using a team-only API key" (presumably
  our integration and staging sending lists don't match our Notify
  account)

Since this is a rather complex web of anticipated errors I've taken the
simple approach of not sending Sentry errors for any Notify exceptions.
These instead will all be written to the log file. Should we find there
is a particular scenario we'd like to hear about quicker we can up the
complexity to be more selective about what is Sentry or not.

I expect this will be complemented by monitoring that determines that a
sufficient number of emails are being sent to Notify which would alert
were we hit to hit a situation where a systemic problem was causing an
unexpectedly high number of emails to fail to send.