Raise errors to trigger retries in DeliveryRequestWorker
Previously when we encountered a transient error in sending an email we
would not retry. The DeliveryAttempt would be marked as a failure and
the Email would remain in a status of pending forever more. This changes
that approach to raise an exception so that the Sidekiq retry system
will try to resend the email.

The error that we raise is excluded from the errors we send to
Sentry. This is because this error represents the application behaving
in an expected way and not something a developer needs to resolve.